🏎️ day-03: way faster by directly selecting the best digits
I particularly love that the Ord monoid approach turned out to be the
fastest in my tests!

It turned out to be even faster than rewriting `maximumBy` to thread
over the first maximum instead of the last:

  maximumBy cmp = foldl1' maxBy
    where
      maxBy x y = case cmp x y of
                    GT -> x
		    EQ -> x
		    LT -> y

I find this surprising. It might be the case that GHC's one which does:

  case cmp x y of
    GT -> x
    _  -> y

has more chances to be fused with the rest of the code?

  <https://github.com/ghc/ghc/blob/fc1d7f7966f56bbe5efaf2796c430dbe526b834b/libraries/ghc-internal/src/GHC/Internal/Data/OldList.hs#L946>

I don't see explicit pragmas about rewriting or fusing but I don't know
what actually *can* be up with this.

Let's just say that `comparing head <> comparing length` is beautiful
and I am very happy it also gives the code a performance boost ☺️