🏎️ day-04: linear solution
The problem is equivalent to the k-core problem described here:

    <https://en.wikipedia.org/wiki/Degeneracy_(graph_theory)>

Which has a O(V+E) linear solution that I implemented in this commit.

We are using a list as a stack because we prepend at most 8 elements,
which means we search depth first and CS.notMember skips re-checks.

The addition of Coord.Set was key to achieve a program runtime that is
comparable to that of an "hello world" Haskell program.

So far all the programs run in the 16-18ms runtime on my machine, this
one being the fastest of the bunch! Very pleased.

I am also liking how every part 1 so far has been super straightforward
and every part 2 had some interesting nuance to it that pays thinking
about. It might be the result of this compressed 25->12 days AoC.